+++ Exception/DeleteException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace XApi\Repository\Api\Exception;
+
+class DeleteException extends WriteException
+{
+}