+++ Exception/SaveException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace XApi\Repository\Api\Exception;
+
+class SaveException extends WriteException
+{
+}