+++ Exception/WriteException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace XApi\Repository\Api\Exception;
+
+/**
+ * @author Jérôme Parmentier <jerome.parmentier@acensi.fr>
+ */
+abstract class WriteException extends \Exception
+{
+}